Minor tweaks
