+++ xarray/test/test_backends.py
@@ -159,11 +159,8 @@ class DatasetIOTestCases(object):
 
             computed = actual.compute()
 
-            # indexes are going to be loaded in memory anyway, even
-            # with compute(). This is by design.
             for v in actual.data_vars.values():
                 self.assertFalse(v._in_memory)
-    
             for v in computed.variables.values():
                 self.assertTrue(v._in_memory)
 
@@ -1021,38 +1018,15 @@ class DaskTest(TestCase, DatasetIOTestCases):
             self.assertEqual(original_names, repeat_names)
 
     def test_dataarray_compute(self):
+        # Test DataArray.compute() on dask backend.
+        # The test for Dataset.compute() is already in DatasetIOTestCases;
+        # however dask is the only tested backend which supports DataArrays
         actual = DataArray([1,2]).chunk()
         computed = actual.compute()
         self.assertFalse(actual._in_memory)
         self.assertTrue(computed._in_memory)
         self.assertDataArrayAllClose(actual, computed)
-
-    def test_pickle(self):
-        # Test that pickling/unpickling does not convert the dask
-        # backend to numpy
-        a1 = DataArray([1,2]).chunk()
-        self.assertFalse(a1._in_memory)        
-        a2 = pickle.loads(pickle.dumps(a1))
-        self.assertDataArrayIdentical(a1, a2)
-        self.assertFalse(a1._in_memory)
-        self.assertFalse(a2._in_memory)        
-
-        ds1 = Dataset({'a': [1,2]}).chunk()
-        self.assertFalse(ds1['a']._in_memory)        
-        ds2 = pickle.loads(pickle.dumps(ds1))
-        self.assertDatasetIdentical(ds1, ds2)
-        self.assertFalse(ds1['a']._in_memory)
-        self.assertFalse(ds2['a']._in_memory)        
-
-    def test_values(self):
-        # Test that invoking the values property does not convert the dask
-        # backend to numpy
-        a = DataArray([1,2]).chunk()
-        self.assertFalse(a._in_memory)
-        self.assertEquals(a.values.tolist(), [1, 2])
-        self.assertFalse(a._in_memory)
-
-
+        
 @requires_scipy_or_netCDF4
 @requires_pydap
 class PydapTest(TestCase):
